mergeClasses prod/non-prod test, ugly solution
